dynamic load, so singular test copy. also test across versions of node, os, etc.
