+++ wasm/packages/okapi/spec/support/jasmine.json
@@ -1,6 +1,6 @@
 {
   "spec_dir": "./test",
-  "spec_files": ["**/*.test.ts"],
+  "spec_files": ["**/*.spec.ts"],
   "helpers": ["./test/helpers/jasmine-helper.ts"],
   "env": {
     "stopSpecOnExpectationFailure": false,
