+++ wasm/packages/okapi/tsconfig.browser.json
@@ -12,6 +12,8 @@
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true
   },
-  "include": ["./src/index.web.ts"],
-  "exclude": ["node_modules", "lib", "test", "./src/index.ts"]
+  "include": [
+    "./src/index.ts"
+  ],
+  "exclude": ["node_modules", "lib", "test"]
 }
